Refactored all primary tests to use new QUnit API
All primary tests (main test.js file) were refactored to use non-global methods (using the 'assert' argument as a namespace) and the new assert.async() method instead of asyncTest. In addition, numerous updates were made to the formatting of the file and there were some updates to how things are run in Grunt.